fix various compile warnings and, since vid_ati_mach64.h is a private header, move it into src/video
