+++ src/video/vid_ati_mach64.h
